GeneticsMakie

Documentation for GeneticsMakie.

GeneticsMakie.calcluateld!

GeneticsMakie.coordinategenes

GeneticsMakie.coordinategwas

GeneticsMakie.coordinateisforms

GeneticsMakie.labelgenome

GeneticsMakie.parsegtf!

GeneticsMakie.plotgenes!

GeneticsMakie.plotgenes!

GeneticsMakie.plotgenes!

GeneticsMakie.plotgwas!

GeneticsMakie.plotisoforms!

GeneticsMakie.plotld!

GeneticsMakie.plotlocus!

GeneticsMakie.plotlocus!

GeneticsMakie.plotlocus!

GeneticsMakie.plotqq!

GeneticsMakie.calcluateld! — Method

calcluateld!(gwas::DataFrame, ref::SnpData; snp)

Calculate LD between the most significant SNP and other SNPs in gwas by using ref. Optionally, the target SNP can be switched to snp.

GeneticsMakie.coordinategenes — Method

coordinategenes(chromosome::AbstractString, range1::Real, range2::Real, gencode::DataFrame, height::Real)

Subset gencode to a given chromosome and genomic range between range1 and range2, and determine coordinates of exons for each gene residing in the genomic region.

GeneticsMakie.coordinategwas — Method

coordinategwas(gwas::Vector{DataFrame}; ymax, chr, bp, p, freey)

Determine shared coordinates for a set of gwas.

GeneticsMakie.coordinateisforms — Method

coordinateisforms(gene::AbstractString, gencode::DataFrame, orderby::Union{Nothing, AbstractVector{<:AbstractString}}, height::Real, text::Bool, textpos::Symbol)

Subset gencode to a gene and determine coordinates of exons for each isoform.

GeneticsMakie.labelgenome — Method

labelgenome(g::GridPosition, chromosome::AbstractString, range1::Real, range2::Real)

GeneticsMakie.parsegtf! — Method

parsegtf!(gencode::DataFrame)

Parse gencode by extracting information from the ninth info column.

GeneticsMakie.plotgenes! — Method

plotgenes!(ax::Axis, gene::AbstractString, gencode::DataFrame; window, height, highlight)

Plot collapsed gene bodies for genes within a certain window around gene. The default window is 1 Mb. Optionally, height of exons can be adjusted using height and genes can be highlighted via highlight.

GeneticsMakie.plotgenes! — Method

plotgenes!(ax::Axis, chromosome::AbstractString, bp::Real, gencode::DataFrame; window, height, highlight)

Plot collapsed gene bodies for genes within a given chromosome and a certain window around a genomic coordinate bp. The default window is 1 Mb. Optionally, height of exons can be adjusted using height and genes can be highlighted via highlight.

GeneticsMakie.plotgenes! — Method

plotgenes!(ax::Axis, chromosome::AbstractString, range1::Real, range2::Real, gencode::DataFrame; height, highlight)

Plot collapsed gene bodies for genes within a given chromosome and genomic range between range1 and range2. Optionally, height of exons can be adjusted using height and genes can be highlighted via highlight.

GeneticsMakie.plotgwas! — Method

plotgwas!(ax::Axis, df::DataFrame, i::Int, ymax::Real, xmax::Real, ticks::DataFrame; xlabel, ylabel, ystep, chr, p, sigline, sigcolor)

Plot gwas results with coordinates from coordinategwas, namely df for phenotype i with x and y limits, xmax and ymax, respectively. The position of x ticks are determined by ticks.

GeneticsMakie.plotisoforms! — Method

plotisoforms!(ax::Axis, gene::AbstractString, gencode::DataFrame; orderby, height, text, textpos)

Plot each isoform of a given gene in a separate row. Optionally, order of isoforms can be changed by orderby, height of exons can be adjusted using height, and the presence or absence of text label of isoforms can be specified using text with its position textpos.

GeneticsMakie.plotld! — Method

plotld!(ax::Axis, LD::AbstractMatrix; color)

Visualize a correlation matrix LD with the diagonal elements on the x-axis.

GeneticsMakie.plotlocus! — Method

plotlocus!(ax::Axis, gene::AbstractString, gwas::DataFrame, gencode::DataFrame; window, colorld, ref, snp, ymax, title)

Plot gwas results within a certain window around gene. The default window is 1 Mb.

GeneticsMakie.plotlocus! — Method

plotlocus!(ax::Axis, chromosome::AbstractString, bp::Real, gwas::DataFrame; window, colorld, ref, snp, ymax, title)

Plot gwas results within a given chromosome and a certain window around a genomic coordinate bp. The default window is 1 Mb.

GeneticsMakie.plotlocus! — Method

plotlocus!(ax::Axis, chromosome::AbstractString, range1::Real, range2::Real, gwas::DataFrame; colorld, ref, snp, ymax)

Plot gwas results within a given chromosome and genomic range between range1 and range2. Optionally, SNPs can be colored by LD via colorld using ref. The default SNP for which LD is calculated is index SNP, which can be changed to snp.

GeneticsMakie.plotqq! — Method

plotqq!(ax::Axis, p::AbstractVector; title, xlabel, ylabel, xstep, ystep)

Plot QQ plot of p values where the expected distribution is the uniform distribution.