API

GeneticsMakie.coordinategenes

GeneticsMakie.coordinategwas

GeneticsMakie.coordinateisforms

GeneticsMakie.findclosestgene

GeneticsMakie.findclosestgene

GeneticsMakie.findgene

GeneticsMakie.findgwasloci

GeneticsMakie.findgwasloci

GeneticsMakie.labelgenome

GeneticsMakie.mungesumstats!

GeneticsMakie.mungesumstats!

GeneticsMakie.parsegtf!

GeneticsMakie.plotgenes!

GeneticsMakie.plotgenes!

GeneticsMakie.plotgenes!

GeneticsMakie.plotgenes!

GeneticsMakie.plotgenes!

GeneticsMakie.plotgenes!

GeneticsMakie.plotgwas!

GeneticsMakie.plotisoforms!

GeneticsMakie.plotld!

GeneticsMakie.plotlocus!

GeneticsMakie.plotlocus!

GeneticsMakie.plotlocus!

GeneticsMakie.plotqq!

GeneticsMakie.coordinategenes — Method

coordinategenes(chromosome::AbstractString, range1::Real, range2::Real, gencode::DataFrame, height::Real)

Subset gencode to a given chromosome and genomic range between range1 and range2, and determine coordinates of exons for each gene residing in the genomic region.

GeneticsMakie.coordinategwas — Method

coordinategwas(gwas::Vector{DataFrame}; freey::Bool)

Determine shared coordinates for a set of gwas.

GeneticsMakie.coordinateisforms — Method

coordinateisforms(gene::AbstractString, gencode::DataFrame, orderby::Union{Nothing, AbstractVector{<:AbstractString}}, height::Real, text::Union{Bool, Symbol})

Subset gencode to a gene and determine coordinates of exons for each isoform.

GeneticsMakie.findclosestgene — Method

findclosestgene(chr::AbstractString, bp::Real, gencode::DataFrame; start::Bool, proteincoding::Bool)

Find the closest gene to a genomic coordinate with chr and bp using gencode. Optionally, the closest gene can be defined from the gene start site considering the strand using start, and only protein coding genes can be considred using proteincoding. The default start and proteincoding are false.

GeneticsMakie.findclosestgene — Method

findclosestgene(df::DataFrame, gencode::DataFrame; kwargs...)

Find the closest genes to each SNP in df using gencode. Optionally, the closest gene can be defined from the gene start site considering the strand using start, and only protein coding genes can be considred using proteincoding. The default start and proteincoding are false.

GeneticsMakie.findgene — Method

findgene(gene::AbstractString, gencode::DataFrame)

Find chromosome, gene start, and gene stop sites for the gene of interest.

GeneticsMakie.findgwasloci — Method

findgwasloci(gwas::DataFrame; p::Real)

Find genome-wide significant loci for gwas based on p threshold that is separated from each other by at least 1 Mb. The default p is 5e-8.

GeneticsMakie.findgwasloci — Method

findgwasloci(gwas::Vector{DataFrame}; p::Real)

Find genome-wide significant loci across multiple gwas based on p threshold that is separated from each other by at least 1 Mb. The default p is 5e-8.

GeneticsMakie.labelgenome — Method

labelgenome(g::GridPosition, chromosome::AbstractString, range1::Real, range2::Real)

GeneticsMakie.mungesumstats! — Method

mungesumstats!(gwas::DataFrame)

GeneticsMakie.mungesumstats! — Method

mungesumstats!(gwas::Vector{DataFrame})

Munge multiple gwas by harmonizing names of columns, their types, and P values, among others. The output contains at least four columns: SNP, CHR, BP, and P corresponding to SNP ID's, chromosome, base pair, and P values.

GeneticsMakie.parsegtf! — Method

parsegtf!(gencode::DataFrame)

Parse gencode by extracting gene_id, gene_name, gene_type, transcript_id, transcript_support_level information from the info column.

GeneticsMakie.plotgenes! — Method

plotgenes!(ax::Axis, gene::AbstractString, gencode::DataFrame; window::Real)

Plot collapsed gene bodies for genes within a certain window around gene. The default window is 1 Mb.

GeneticsMakie.plotgenes! — Method

plotgenes!(ax::Axis, chromosome::AbstractString, bp::Real, gencode::DataFrame; window::Real)

Plot collapsed gene bodies for genes within a given chromosome and a certain window around a genomic coordinate bp. The default window is 1 Mb.

GeneticsMakie.plotgenes! — Method

plotgenes!(ax::Axis, chromosome::AbstractString, range1::Real, range2::Real, gencode::DataFrame; height::Real, genecolor, textcolor)

Plot collapsed gene bodies for genes within a given chromosome and genomic range between range1 and range2. Optionally, the height of exons can be adjusted using height and color of genes or gene names adjusted using genecolor and textcolor, respectively.

GeneticsMakie.plotgenes! — Method

plotgenes!(ax::Axis, chromosome::AbstractString, range1::Real, highlight::Tuple{AbstractVector, AbstractVector}, range2::Real, gencode::DataFrame; height::Real)

Plot collapsed gene bodies for genes within a given chromosome and genomic range between range1 and range2. Optionally, the height of exons can be adjusted using height. Genes to be highlighted and their colors can be specified via highlight.

GeneticsMakie.plotgenes! — Method

plotgenes!(ax::Axis, chromosome::AbstractString, bp::Real, highlight::Tuple{AbstractVector, AbstractVector}, gencode::DataFrame; window::Real)

Plot collapsed gene bodies for genes within a given chromosome and a certain window around a genomic coordinate bp. The default window is 1 Mb. Genes to be highlighted and their colors can be specified via highlight.

GeneticsMakie.plotgenes! — Method

plotgenes!(ax::Axis, gene::AbstractString, highlight::Tuple{AbstractVector, AbstractVector}, gencode::DataFrame; window::Real)

Plot collapsed gene bodies for genes within a certain window around gene. The default window is 1 Mb. Genes to be highlighted and their colors can be specified via highlight.

GeneticsMakie.plotgwas! — Method

plotgwas!(ax::Axis, df::DataFrame, i::Int, ymax::Real, xmax::Real, ticks::DataFrame; ystep::Real, sigline::Bool, sigcolor::Bool)

Plot gwas results with coordinates from coordinategwas, namely df for phenotype i with x and y limits, xmax and ymax, respectively. The position of x ticks are determined by ticks.

GeneticsMakie.plotisoforms! — Method

plotisoforms!(ax::Axis, gene::AbstractString, gencode::DataFrame; orderby::Union{Nothing, AbstractVector{<:AbstractString}}, height::Real, isoformcolor, textcolor, text::Union{Bool, Symbol})

Plot each isoform of a given gene in a separate row. Optionally, the order of isoforms can be changed by orderby, height of exons can be adjusted using height, and color of isoforms or isoform names adjusted using isoformcolor and textcolor, respectively. The position of text label of isoforms can be specified using text.

GeneticsMakie.plotld! — Method

plotld!(ax::Axis, LD::AbstractMatrix; color::AbstractString)

Visualize a symmetric correlation matrix LD with the diagonal elements on the x-axis.

GeneticsMakie.plotlocus! — Method

plotlocus!(ax::Axis, gene::AbstractString, gwas::DataFrame, gencode::DataFrame; window::Real, kwargs...)

Plot gwas results within a certain window around gene. The default window is 1 Mb.

GeneticsMakie.plotlocus! — Method

plotlocus!(ax::Axis, chromosome::AbstractString, bp::Real, gwas::DataFrame; window::Real, kwargs...)

Plot gwas results within a given chromosome and a certain window around a genomic coordinate bp. The default window is 1 Mb.

GeneticsMakie.plotlocus! — Method

plotlocus!(ax::Axis, chromosome::AbstractString, range1::Real, range2::Real, gwas::DataFrame; ld::Union{Bool, SnpData, Tuple{SnpData, Union{AbstractString, Tuple{AbstractString, Int}}}}, ymax::Real)

Plot gwas results within a given chromosome and genomic range between range1 and range2. Optionally, SNPs can be colored by LD via ld. The default SNP for which LD is calculated is index SNP.

GeneticsMakie.plotqq! — Method

plotqq!(ax::Axis, P::AbstractVector; xstep::Real, ystep::Real)

Plot QQ plot of P values where the expected distribution is the uniform distribution.